requirements.txt
flask
supabase
python-dotenv
gunicorn
google-generativeai
sentence-transformers
torch